Backend/requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dateutil==2.8.2
python-decouple==3.8
redis==5.0.1
requests==2.25.1
SecretStorage==3.3.1
firebase-admin==6.5.0
phonenumbers==8.13.37
Pillow==10.3.0 
shapely==2.0.4 
twilio==9.2.2 
scipy==1.14.0 
geopy==2.4.1 